- refs #81965: Moved Frequency dropdown to its own component so it can be reused in both sections.
